If the database does not exist then the server should create an empty one.  I already had the code for this, it just needed moving around.

+++ sys/devutils/GetSchema.js
@@ -1,50 +1,15 @@
 var fs = require('fs');
 var win32ole = require('../server/node_modules/win32ole');
-
-// Define the ADO constants that we need
-
-// SchemaEnumConstants
-var adSchemaTables = 20;
-
-// CursorTypeEnum constants
-var adOpenStatic = 3;
-var adOpenKeyset = 1;
-
-// LockTypeEnum constants
-var adLockReadOnly = 1;
-var adLockOptimistic = 3;
-
-// CommandTypeEnum constants
-var adCmdTable = 2;
-var adCmdTableDirect = 512;
-
-// DataTypeEnum constants
-var adInteger  = 3;
-var adDate     = 7;
-var adBoolean  = 11;
-var adVarWChar = 202;
-var adLongVarWChar = 203;
-
-function openAccessDatabase(sFilename)
-{
-    var db = win32ole.client.Dispatch('ADODB.Connection');
-    db.Provider = "Microsoft.Jet.OLEDB.4.0";
-    db.Open(sFilename);
-    return db;
-}
-
-function createRecordset()
-{
-    return win32ole.client.Dispatch('ADODB.Recordset');
-}
+var platform = require('../server/usingMSJet4.js');
+var jet = platform.jet;
 
 function getSchema(sFilename)
 {
     var sSql = "";
-    
-    var db = openAccessDatabase(sFilename);
 
-    var TablesSchema = db.OpenSchema(adSchemaTables);
+    var db = jet.openAccessDatabase(sFilename);
+
+    var TablesSchema = db.OpenSchema(jet.adSchemaTables);
     while (TablesSchema.EOF == false)
     {
         var sTablename = String(TablesSchema.Fields("TABLE_NAME").Value);
@@ -64,8 +29,8 @@ function getSchema(sFilename)
 function getTransportManagerSchema(sFilename)
 {
     var sSql = "";
-    
-    var db = openAccessDatabase(sFilename);
+
+    var db = jet.openAccessDatabase(sFilename);
     sSql += getTable(db, "Clients");
     sSql += getTable(db, "Destinations");
     sSql += getTable(db, "DestinationType");
@@ -82,8 +47,8 @@ function getTable(db, sTablename)
 {
     var sSql = "";
 
-    var rs = createRecordset();
-    rs.Open(sTablename, db, adOpenStatic, adLockReadOnly, adCmdTableDirect);
+    var rs = jet.createRecordset();
+    rs.Open(sTablename, db, jet.adOpenStatic, jet.adLockReadOnly, jet.adCmdTableDirect);
 
     sSql += "\r\n";
     sSql += "CREATE TABLE " + sTablename + "(\r\n";
@@ -97,23 +62,23 @@ function getTable(db, sTablename)
 
         switch (typeInfo)
         {
-        case adInteger:
+        case jet.adInteger:
             typeInfo = "INTEGER";
             break;
 
-        case adBoolean:
+        case jet.adBoolean:
             typeInfo = "YESNO";
             break;
 
-        case adVarWChar:
+        case jet.adVarWChar:
             typeInfo = "TEXT(" + fld.DefinedSize + ")";
             break;
 
-        case adLongVarWChar:
+        case jet.adLongVarWChar:
             typeInfo = "MEMO";
             break;
 
-        case adDate:
+        case jet.adDate:
             typeInfo = "DATE";
             break;
         }
@@ -130,68 +95,22 @@ function getTable(db, sTablename)
     return sSql;
 }
 
-function copyFile(source, target, doneCopying)
+var sFilenameExisting = "./TransportManager.mdb";
+var sFilenameSql = "../TransportManager.sql";
+var sFilenameClone = "../../TransportManager.mdb"
+
+//var sSql = getSchema(sFilenameExisting);
+var sSql = getTransportManagerSchema(sFilenameExisting);
+
+fs.writeFile(sFilenameSql, sSql);
+
+// If the main database is present, get it's schema too so we can check they're the same
+fs.exists(sFilenameClone, function(bExists)
 {
-    var streamIn = fs.createReadStream(source);
-    streamIn.on("error", function(err)
+    if (bExists)
     {
-        console.log(err);
-    });
-    
-    var streamOut = fs.createWriteStream(target);
-    streamOut.on("error", function(err)
-    {
-        console.log(err);
-    });
-
-    streamOut.on("close", function(ex)
-    {
-        doneCopying();
-    });
-
-    streamIn.pipe(streamOut);
-}
-
-function createEmptyClone(sFilenameExisting, sFilenameClone, sFilenameEmpty, sFilenameSql, doneCloning)
-{
-//    var sSql = getSchema(sFilenameExisting);
-    var sSql = getTransportManagerSchema(sFilenameExisting);
-
-    fs.writeFile(sFilenameSql, sSql);
-    
-    // Ensure database doesn't already exist
-    if (fs.existsSync(sFilenameClone))
-    {
-        fs.unlinkSync(sFilenameClone);
-    }
-
-    // Copy empty database
-    copyFile(sFilenameEmpty, sFilenameClone, function()
-    {
-        // Run the SQL on the empty database
-        var x = 0;
-        var db = openAccessDatabase(sFilenameClone);
-
-        arrSql = sSql.split(";");
-        for (var n = 0; n < arrSql.length - 1; ++n)
-        {
-            var statement = arrSql[n];
-            db.Execute(statement);
-        }
-        
-        db.Close();
-
-        // Finally get the schema of the newly cloned database so we can check they're the same
-        sSql = getTransportManagerSchema(sFilenameExisting);
+        sSql = getTransportManagerSchema(sFilenameClone);
         fs.writeFile(sFilenameSql + "2", sSql);
+    }
+});
 
-        doneCloning();
-    });
-}
-
-
-createEmptyClone('./TransportManager.mdb',           // sFilenameExisting
-                 '../../TransportManager.mdb',       // sFilenameClone
-                 '../server/Blank2002Database.mdb',  // sFilenameEmpty
-                 '../TransportManager.sql',          // sFilenameSql
-                 function() { });                    // doneCloning()
